requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
bio-attention==0.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
deepsmiles==1.0.1
easydev==0.13.3
einops==0.8.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
grpcio==1.69.0
h5py==3.12.1
h5torch==0.2.14
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
lightning==2.5.0.post0
lightning-utilities==0.11.9
line_profiler==4.2.0
llvmlite==0.43.0
maldi-nn==0.2.5
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
ptyprocess==0.7.0
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
rdkit==2024.9.4
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selfies==2.2.0
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.5.1
torchmetrics==1.6.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
Werkzeug==3.1.3
xgboost==2.1.3
yarl==1.18.3
